+++ modules/drivers/virtio/pci.c
@@ -152,7 +152,7 @@ virtio_driver_t *virtio_pci_init(void *data) {
         uint64_t bar_paddr = bar->address + pci->common_cfg->offset;
         if (bar_paddr == 0)
             goto done;
-        uint64_t bar_vaddr = phys_to_virt(bar_paddr);
+        uint64_t bar_vaddr = (uint64_t)phys_to_virt(bar_paddr);
         map_page_range(get_current_page_dir(false), bar_vaddr, bar_paddr,
                        pci->common_cfg->length,
                        PT_FLAG_R | PT_FLAG_W | PT_FLAG_DEVICE);
@@ -175,7 +175,7 @@ virtio_driver_t *virtio_pci_init(void *data) {
         uint64_t bar_paddr = bar->address + notify_cfg->offset;
         if (bar_paddr == 0)
             goto done;
-        uint64_t bar_vaddr = phys_to_virt(bar_paddr);
+        uint64_t bar_vaddr = (uint64_t)phys_to_virt(bar_paddr);
         map_page_range(get_current_page_dir(false), bar_vaddr, bar_paddr,
                        notify_cfg->length,
                        PT_FLAG_R | PT_FLAG_W | PT_FLAG_DEVICE);
@@ -198,7 +198,7 @@ virtio_driver_t *virtio_pci_init(void *data) {
         uint64_t bar_paddr = bar->address + pci->device_cfg->offset;
         if (bar_paddr == 0)
             goto done;
-        uint64_t bar_vaddr = phys_to_virt(bar_paddr);
+        uint64_t bar_vaddr = (uint64_t)phys_to_virt(bar_paddr);
         map_page_range(get_current_page_dir(false), bar_vaddr, bar_paddr,
                        pci->device_cfg->length,
                        PT_FLAG_R | PT_FLAG_W | PT_FLAG_DEVICE);
